# docs/api-contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: AirServiz API – Sprint 1 Contracts
  description: |
    Contracts for the core AirServiz microservices exposed via the Kong API Gateway.
    All endpoints are prefixed by Kong at `/api/v1`.
    Authentication uses Bearer JWT tokens issued by the `auth-service`.
  version: "1.0.0"
  contact:
    name: AirServiz Engineering
  license:
    name: MIT

servers:
  - url: http://localhost:8000/api/v1
    description: Local – Kong Gateway (dev)
  - url: https://api.airserviz.com/v1
    description: Production

# ─── Security ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable Schemas ───────────────────────────────────────────────────────
  schemas:

    # ── Common ──
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer, example: 400 }
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error: { type: string, example: "Bad Request" }

    UUID:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"

    # ── Auth ──
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@airserviz.com"
        password:
          type: string
          minLength: 8
          example: "S3cur3P@ss!"

    TokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:  { type: string, description: "JWT – 15 min TTL" }
        refreshToken: { type: string, description: "Opaque token – 7 day TTL" }
        expiresIn:    { type: integer, example: 900, description: "Seconds" }

    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }

    # ── Order ──
    OrderItem:
      type: object
      required: [serviceId, quantity, unitPrice]
      properties:
        serviceId:
          $ref: "#/components/schemas/UUID"
        quantity:
          type: integer
          minimum: 1
          example: 2
        unitPrice:
          type: number
          format: double
          minimum: 0.01
          example: 150.00

    CreateOrderRequest:
      type: object
      required: [providerId, items]
      properties:
        providerId:
          $ref: "#/components/schemas/UUID"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: string
          maxLength: 500
          example: "Please arrive before 9 AM"

    Order:
      type: object
      required: [id, clientId, providerId, items, status, totalAmount, createdAt]
      properties:
        id:           { $ref: "#/components/schemas/UUID" }
        clientId:     { $ref: "#/components/schemas/UUID" }
        providerId:   { $ref: "#/components/schemas/UUID" }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        notes:        { type: string }
        status:
          type: string
          enum: [PENDING, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED]
          example: PENDING
        totalAmount:
          type: number
          format: double
          example: 300.00
        createdAt:
          type: string
          format: date-time

    OrderCreatedEvent:
      description: |
        Kafka event emitted to topic `order_created` after a successful POST /orders.
      type: object
      required: [eventType, version, payload, metadata]
      properties:
        eventType:  { type: string, enum: [order_created] }
        version:    { type: string, example: "1.0" }
        payload:    { $ref: "#/components/schemas/Order" }
        metadata:
          type: object
          required: [timestamp, correlationId]
          properties:
            timestamp:     { type: string, format: date-time }
            correlationId: { $ref: "#/components/schemas/UUID" }

    # ── User ──
    UserProfile:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id:        { $ref: "#/components/schemas/UUID" }
        email:     { type: string, format: email }
        fullName:  { type: string, example: "María González" }
        role:
          type: string
          enum: [CLIENT, PROVIDER, ADMIN]
          example: CLIENT
        avatarUrl: { type: string, format: uri }
        createdAt: { type: string, format: date-time }

# ─── Global security ─────────────────────────────────────────────────────────
security:
  - bearerAuth: []

# ─── Tags ────────────────────────────────────────────────────────────────────
tags:
  - name: auth
    description: "Authentication – JWT issuance and refresh (auth-service)"
  - name: orders
    description: "Order lifecycle and Kafka event emission (booking-service)"
  - name: users
    description: "User profile management (user-service)"

# ─── Paths ───────────────────────────────────────────────────────────────────
paths:

  # ── 1. POST /auth/login ────────────────────────────────────────────────────
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate and obtain JWT tokens
      description: |
        Validates credentials and returns an `accessToken` (JWT, 15 min) and
        a `refreshToken` (7 days). The access token must be sent as
        `Authorization: Bearer <token>` on all subsequent requests.
      security: []   # public endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "user@airserviz.com"
              password: "S3cur3P@ss!"
      responses:
        "200":
          description: Tokens issued successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── 2. POST /auth/refresh ──────────────────────────────────────────────────
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token using a valid refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Refresh token expired or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── 3. POST /orders ────────────────────────────────────────────────────────
  /orders:
    post:
      tags: [orders]
      summary: Create a new order and emit order_created Kafka event
      description: |
        Accepts an order payload, creates an `Order` entity with status `PENDING`,
        and immediately emits an `order_created` event to the Kafka topic
        `order_created`. The `x-client-id` header is injected automatically
        by the Kong JWT plugin in production environments.
      parameters:
        - name: x-client-id
          in: header
          required: true
          description: Authenticated client UUID (injected by Kong in production)
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
            example:
              providerId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
              items:
                - serviceId: "7b9e2a1d-3c4f-5e6a-7b8c-9d0e1f2a3b4c"
                  quantity: 1
                  unitPrice: 150.00
              notes: "Please bring cleaning supplies"
      responses:
        "201":
          description: Order created — Kafka event emitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: "#/components/schemas/Order" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── 4. GET /users/me ───────────────────────────────────────────────────────
  /users/me:
    get:
      tags: [users]
      summary: Get the authenticated user's profile
      description: Returns the profile of the user identified by the Bearer JWT.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── 5. GET /orders/{id} ────────────────────────────────────────────────────
  /orders/{id}:
    get:
      tags: [orders]
      summary: Get order details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:    { $ref: "#/components/schemas/Order" }
        "404":
          description: Order not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }